pyproject.toml
[tool.poetry]
name = "cmus_buddy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
rich = "^10.16.1"
imgcat = "^0.5.0"
Pillow = "^8.4.0"
libtmux = "^0.10.2"
python3-discogs-client = "^2.3.12"
py-bandcamp = "^0.7.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
